backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-ai==0.0.16
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

